Time: 27 ms (23.65%), Space: 47.8 MB (28.89%) - LeetHub

+++ 0227-basic-calculator-ii/0227-basic-calculator-ii.java
@@ -2,7 +2,7 @@ class Solution {
     public int calculate(String s) {
         Stack<Integer> st = new Stack<>();
         char c = '+';
-        long n = 0;
+        int  n = 0;
 
         for (int i = 0; i < s.length(); i++) {
             char c1 = s.charAt(i);
